Split AO.py into atomic mixing files
